+++ src/Journalist.WindowsAzure.Storage.Abstractions/Queues/ICloudQueue.cs
@@ -16,7 +16,11 @@ namespace Journalist.WindowsAzure.Storage.Queues
 
         Task<IReadOnlyList<ICloudQueueMessage>> GetMessagesAsync();
 
+        Task<IReadOnlyList<ICloudQueueMessage>> GetMessagesAsync(int messageCount);
+
         Task<IReadOnlyList<ICloudQueueMessage>> GetMessagesAsync(TimeSpan visibilityTimeout);
+
+        Task<IReadOnlyList<ICloudQueueMessage>> GetMessagesAsync(int messageCount, TimeSpan visibilityTimeout);
 
         Task UpdateMessageAsync(string messageId, string popReceipt, byte[] content, TimeSpan visibilityTimeout);