+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ### 0.8.4
 * [WAZ.Storage][New] CloudQueue getting messages operation accept visibility timeout parameter.
+* [WAZ.Storage][New] CloudQueue getting messages operation accept message count parameter.
 * [EventStore][Improvment] Consumer can start reading from the end of streams.
 * [EventStore][Breaking] Notification listener starts reading stream from the end.
 * [EventStore][BugFix] Consumer initialize reader only in leader state.